bulk refactoring

+++ src/controllers/users/delete.py
@@ -1,20 +1,26 @@
 import db.conn
 import db.models
-from controllers.helper import transaction, api_result
+from controllers.helper import transaction, api_result, token_to_auth_user
+from journaling import log
+from controllers.models import AuthUser, UNAUTH_OPER_CODE, APP_ERROR_CODE
 
 
 @api_result
 @transaction
-def delete_user(user_id: str):
+@token_to_auth_user
+def delete_user(auth_user: AuthUser, user_id: str):
     """
     Deletes user.
     Returns deleted user id.
     """
+    if not auth_user.is_admin:
+        return 'Only admin can delete users', UNAUTH_OPER_CODE
     user_to_delete = db.conn.session.query(db.models.User).filter(db.models.User.id == user_id)
     if user_to_delete.count():
+            log.debug(f'Deletion of user with id={user_id}')
             db.conn.session.delete(user_to_delete.first())
             db.conn.session.commit()
             return None
     else:
-        return f'No user with id={user_id}', 400
-
+        log.debug(f'No user with id={user_id}')
+        return f'No user with id={user_id}', APP_ERROR_CODE
